+++ cfdm/core/__init__.py
@@ -5,7 +5,7 @@
 __author__ = 'David Hassell'
 __date__ = '2020-06-10'
 __cf_version__ = '1.8'
-__version__ = '1.8.5'
+__version__ = '1.8.6'
 
 from distutils.version import LooseVersion
 import platform