+++ cfdm/mixin/netcdf.py
@@ -595,7 +595,7 @@ class NetCDFGlobalAttributes(NetCDF):
     :Returns:
 
         `dict`
-            The selection of properties requested for writting to
+            The selection of properties requested for writing to
             netCDF global attributes.
 
     **Examples:**
@@ -660,7 +660,7 @@ class NetCDFGlobalAttributes(NetCDF):
     :Returns:
 
         `dict`
-            The removed selection of properties requested for writting
+            The removed selection of properties requested for writing
             to netCDF global attributes.
 
     **Examples:**
@@ -832,28 +832,42 @@ class NetCDFGlobalAttributes(NetCDF):
 class NetCDFGroups(NetCDF):
     '''Mixin class for accessing netCDF groups.
     
-    Classes which inherit from this class must also inherit from
-    `NetCDFVariable`
+    Classes which inherit from this class must also inherit from the
+    `NetCDFVariable` mixin class.
 
     .. versionadded:: 1.8.6
 
     '''
     def nc_groups(self):
-        '''TODO Return the netCDF variable name.
+        '''Return the netCDF group hierarchy.
 
+    The group hierarchy is defined by the netCDF variable name. Groups
+    are delimited by / (slash) characters in the netCDF variable
+    name. The groups are returned, in hierarchical order, as a
+    sequence of strings. If the netCDF variable is not set, or
+    contains no / characters then an empty sequence is returned,
+    signifying the root group.
+    
     .. versionadded:: 1.8.6
 
-    .. seealso:: TODO
-
+    .. seealso:: `nc_get_variable`, `nc_set_variable`
+    
     :Returns:
 
-        `tuple`
-            TODO The netCDF variable name. If unset then *default* is
-            returned, if provided.
+        `tuple` of `str`
+            The netCDF group names. An empty tuple signifies the root
+            group.
 
     **Examples:**
 
-TODO
+    >>> f.nc_get_variable()
+    'q'
+    >>> nc.groups()
+    ()
+        
+    >>> f.nc_set_variable('/forecast/model/q')
+    >>> nc.groups()
+    ('forecast', 'model')
 
         '''
         ncvar = self.nc_get_variable(None)
@@ -861,7 +875,8 @@ TODO
             return ()
 
         return tuple(ncvar.split('/')[1:-1])
-        
+
+    
 class NetCDFGroupAttributes(NetCDF):
     '''Mixin class for accessing netCDF group attributes.
     
@@ -869,58 +884,42 @@ class NetCDFGroupAttributes(NetCDF):
 
     '''
     def nc_group_attributes(self, values=False):
-        '''Return the selection of properties to be written as netCDF4 group
+        '''Return the selection of properties to be written as netCDF group
     attributes.
-
-TODO
-    When multiple field constructs are being written to the same file,
-    it is only possible to create a netCDF global attribute from a
-    property that has identical values for each field construct. If
-    any field construct's property has a different value then the
-    property will not be written as a netCDF global attribute, even if
-    it has been selected as such, but will appear instead as
-    attributes on the netCDF data variables corresponding to each
-    field construct.
-
-    The standard description-of-file-contents properties are always
-    written as netCDF global attributes, if possible, so selecting
-    them is optional.
 
     .. versionadded:: 1.8.6
 
-    .. seealso:: `write`, `nc_clear_global_attributes`,
-                 `nc_set_global_attribute`, `nc_set_global_attributes`
+    .. seealso:: `write`, `nc_clear_group_attributes`,
+                 `nc_set_group_attribute`, `nc_set_group_attributes`
 
     :Parameters:
 
         values: `bool`, optional
-            Return the value (rather than `None`) for any global
+            Return the value (rather than `None`) for any group
             attribute that has, by definition, the same value as a
             construct property.
-
-            .. versionadded:: 1.8.2
 
     :Returns:
 
         `dict`
-            The selection of properties requested for writting to
-            netCDF global attributes.
+            The selection of properties requested for writing to
+            netCDF group attributes.
 
     **Examples:**
 
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None}
-    >>> f.nc_set_global_attribute('foo')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None, 'foo': None}
-    >>> f.nc_set_global_attribute('comment', 'global comment')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': None}
-    >>> f.nc_global_attributes(values=True)
-    {'Conventions': 'CF-1.8', 'comment': 'global_comment', 'foo': 'bar'}
-    >>> f.nc_clear_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': None}
-    >>> f.nc_global_attributes()
+    >>> f.nc_group_attributes()
+    {'comment': None}
+    >>> f.nc_set_group_attribute('foo')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': None}
+    >>> f.nc_set_group_attribute('foo', 'bar')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes(values=True)
+    {'comment': 'forecast comment', 'foo': 'bar'}
+    >>> f.nc_clear_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes()
     {}
 
         '''
@@ -944,47 +943,35 @@ TODO
         return out
 
     def nc_clear_group_attributes(self):
-        '''Remove the selection of properties to be written as netCDF4 group
+        '''Remove the selection of properties to be written as netCDF group
     attributes.
 
-TODO
-    When multiple field constructs are being written to the same file,
-    it is only possible to create a netCDF global attribute from a
-    property that has identical values for each field construct. If
-    any field construct's property has a different value then the
-    property will not be written as a netCDF global attribute, even if
-    it has been selected as such, but will appear instead as
-    attributes on the netCDF data variables corresponding to each
-    field construct.
+    .. versionadded:: 1.8.6
 
-    The standard description-of-file-contents properties are always
-    written as netCDF global attributes, if possible, so selecting
-    them is optional.
-
-    .. versionadded:: 1.7.0
-
-    .. seealso:: `write`, `nc_global_attributes`,
-                 `nc_set_global_attribute`, `nc_set_global_attributes`
+    .. seealso:: `write`, `nc_group_attributes`,
+                 `nc_set_group_attribute`, `nc_set_group_attributes`
 
     :Returns:
 
         `dict`
-            The removed selection of properties requested for writting
-            to netCDF global attributes.
+            The removed selection of properties requested for writing
+            to netCDF group attributes.
 
     **Examples:**
 
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None}
-    >>> f.nc_set_global_attribute('foo')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None, 'foo': None}
-    >>> f.nc_set_global_attribute('comment', 'global comment')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': None}
-    >>> f.nc_clear_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': None}
-    >>> f.nc_global_attributes()
+    >>> f.nc_group_attributes()
+    {'comment': None}
+    >>> f.nc_set_group_attribute('foo')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': None}
+    >>> f.nc_set_group_attribute('foo', 'bar')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes(values=True)
+    {'comment': 'forecast comment', 'foo': 'bar'}
+    >>> f.nc_clear_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes()
     {}
 
         '''
@@ -998,40 +985,26 @@ TODO
         return out
 
     def nc_set_group_attribute(self, prop, value=None):
-        '''Select a property to be written as a netCDF4 group attribute.
+        '''Select a property to be written as a netCDF group attribute.
 
-TODO
-    When multiple field constructs are being written to the same file,
-    it is only possible to create a netCDF global attribute from a
-    property that has identical values for each field construct. If
-    any field construct's property has a different value then the
-    property will not be written as a netCDF global attribute, even if
-    it has been selected as such, but will appear instead as
-    attributes on the netCDF data variables corresponding to each
-    field construct.
+    .. versionadded:: 1.8.6
 
-    The standard description-of-file-contents properties are always
-    written as netCDF global attributes, if possible, so selecting
-    them is optional.
-
-    .. versionadded:: 1.7.0
-
-    .. seealso:: `write`, `nc_global_attributes`,
-                 `nc_clear_global_attributes`,
-                 `nc_set_global_attributes`
+    .. seealso:: `write`, `nc_group_attributes`,
+                 `nc_clear_group_attributes`,
+                 `nc_set_group_attributes`
 
     :Parameters:
 
         prop: `str`
             Select the property to be written (if possible) as a
-            netCDF global attribute.
+            netCDF group attribute.
 
         value: optional
-            The value of the netCDF global attribute, which will be
+            The value of the netCDF group attribute, which will be
             created (if possible) in addition to the property as
             written to a netCDF data variable. If unset (or `None`)
             then this acts as an instruction to write the property (if
-            possible) to a netCDF global attribute instead of to a
+            possible) to a netCDF group attribute instead of to a
             netCDF data variable.
 
     :Returns:
@@ -1040,17 +1013,19 @@ TODO
 
     **Examples:**
 
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None}
-    >>> f.nc_set_global_attribute('foo')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None, 'foo': None}
-    >>> f.nc_set_global_attribute('comment', 'global comment')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': None}
-    >>> f.nc_clear_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': None}
-    >>> f.nc_global_attributes()
+    >>> f.nc_group_attributes()
+    {'comment': None}
+    >>> f.nc_set_group_attribute('foo')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': None}
+    >>> f.nc_set_group_attribute('foo', 'bar')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes(values=True)
+    {'comment': 'forecast comment', 'foo': 'bar'}
+    >>> f.nc_clear_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes()
     {}
 
         '''
@@ -1064,37 +1039,22 @@ TODO
         self._get_component('netcdf')['group_attributes'] = out
 
     def nc_set_group_attributes(self, properties, copy=True):
-        '''Set properties to be written as netCDF4 group attributes.
+        '''Set properties to be written as netCDF group attributes.
 
-TODO
+    .. versionadded:: 1.8.6
 
-    When multiple field constructs are being written to the same file,
-    it is only possible to create a netCDF global attribute from a
-    property that has identical values for each field construct. If
-    any field construct's property has a different value then the
-    property will not be written as a netCDF global attribute, even if
-    it has been selected as such, but will appear instead as
-    attributes on the netCDF data variables corresponding to each
-    field construct.
-
-    The standard description-of-file-contents properties are always
-    written as netCDF global attributes, if possible, so selecting
-    them is optional.
-
-    .. versionadded:: 1.7.10
-
-    .. seealso:: `write`, `nc_clear_global_attributes`,
-                 `nc_global_attributes`, `nc_set_global_attribute`
+    .. seealso:: `write`, `nc_clear_group_attributes`,
+                 `nc_group_attributes`, `nc_set_group_attribute`
 
     :Parameters:
 
         properties: `dict`
-            Set the properties be written as a netCDF global attribute
-            from the dictionary supplied. The value of a netCDF global
+            Set the properties be written as a netCDF group attribute
+            from the dictionary supplied. The value of a netCDF group
             attribute, which will be created (if possible) in addition
             to the property as written to a netCDF data variable. If a
             value of `None` is used then this acts as an instruction
-            to write the property (if possible) to a netCDF global
+            to write the property (if possible) to a netCDF group
             attribute instead of to a netCDF data variable.
 
             *Parameter example:*
@@ -1111,18 +1071,20 @@ TODO
 
     **Examples:**
 
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None}
-    >>> f.nc_set_global_attributes({})
-    >>> f.nc_set_global_attributes({'foo': None})
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': None, 'foo': None}
-    >>> f.nc_set_global_attributes('comment', 'global comment')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': None}
-    >>> f.nc_set_global_attributes('foo', 'bar')
-    >>> f.nc_global_attributes()
-    {'Conventions': None, 'comment': 'global_comment', 'foo': 'bar'}
+    >>> f.nc_group_attributes()
+    {'comment': None}
+    >>> f.nc_set_group_attribute('foo')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': None}
+    >>> f.nc_set_group_attribute('foo', 'bar')
+    >>> f.nc_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes(values=True)
+    {'comment': 'forecast comment', 'foo': 'bar'}
+    >>> f.nc_clear_group_attributes()
+    {'comment': None, 'foo': 'bar'}
+    >>> f.nc_group_attributes()
+    {}
 
         '''
         if copy: