+++ Changelog.rst
@@ -1,3 +1,19 @@
+version 1.8.6
+-------------
+----
+
+**2020-07-??**
+
+* Implemented reading and writing netCDF4 group hierarchies for CF-1.8
+  (https://github.com/NCAS-CMS/cfdm/issues/13)
+* New method: `cfdm.Field.nc_groups`
+* New method: `cfdm.Field.nc_group_attributes`
+* New method: `cfdm.Field.nc_set_group_attribute`
+* New method: `cfdm.Field.nc_set_group_attributes`
+* New method: `cfdm.Field.nc_clear_group_attributes`
+* New keyword parameter to `cfdm.write`: ``group``
+* new dependency: ``netcdf_flattener>=1.1.0``
+
 version 1.8.5
 -------------
 ----